Prometheus documentation: default to listening on all addresses.
`localhost:9430` picks one particular IP/interface, which can be quite
confusing on systems with many interfaces or addresses. It likely makes
most sense to default to listening on all addresses so this does not
cause confusion.

(And we expect users to have adequate firewalling to prevent access from
WAN if relevant)